Refactorización para seguimiento de buenas prácticas y que esté todo limpio

+++ src/main/java/cl/javadevs/springsecurityjwt/services/ServicioService.java
@@ -1,5 +1,6 @@
 package cl.javadevs.springsecurityjwt.services;
 
+import cl.javadevs.springsecurityjwt.exceptions.ServicioNoEncontradoException;
 import cl.javadevs.springsecurityjwt.models.Servicio;
 import cl.javadevs.springsecurityjwt.repositories.IServicioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,7 +30,8 @@ public class ServicioService {
 
     //Obtenemos un celular por su id
     public Optional<Servicio> readOne(Long id) {
-        return servicioRepo.findById(id);
+        return Optional.ofNullable(servicioRepo.findById(id)
+                .orElseThrow(() -> new ServicioNoEncontradoException("Servicio no encontrado con Id: "+id)));
     }
 
     //Actualizamos un celular
@@ -39,6 +41,9 @@ public class ServicioService {
 
     //Eliminamos un celular
     public void delete(Long id) {
+        if (!servicioRepo.existsById(id)) {
+            throw new ServicioNoEncontradoException("No se puede eliminar. Servicio no encontrado con Id: " + id);
+        }
         servicioRepo.deleteById(id);
     }
 }
